% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_basis_additive.R
\name{GetDoF.addHierBasis}
\alias{GetDoF.addHierBasis}
\title{Degrees of Freedom for Additive HierBasis}
\usage{
GetDoF.addHierBasis(object)
}
\arguments{
\item{object}{An object of class \code{addHierBasis}.}
}
\value{
A numeric vector of degrees of freedom for the sequence of lambda values
}
\description{
Degrees of Freedom for Additive HierBasis
}
\examples{

library(HierBasis)
require(Matrix)

set.seed(1)

# Generate the points x.
n <- 100
p <- 30

x <- matrix(rnorm(n*p), ncol = p)

# A simple model with 3 non-zero functions.
y <- rnorm(n, sd = 0.1) + sin(x[, 1]) + x[, 2] + (x[, 3])^3

mod <- AdditiveHierBasis(x, y, nbasis = 50, max.lambda = 30,
                         beta.mat = NULL,
                         nlam = 50, alpha = 0.5,
                         lam.min.ratio = 1e-4, m.const = 3,
                         max.iter = 300, tol = 1e-4)
df <- GetDoF.addHierBasis(mod)
}
\references{
Haris, A., Shojaie, A. and Simon, N. (2018). Nonparametric Regression with
Adaptive Smoothness via a Convex Hierarchical Penalty. Available on request
by authors.
}
\seealso{
\code{\link{AdditiveHierBasis}}
}
\author{
Asad Haris (\email{aharis@uw.edu}),
Ali Shojaie and Noah Simon
}
