% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_basis_additive.R
\name{plot.addHierBasis}
\alias{plot.addHierBasis}
\title{Plot function for \code{addHierBasis}}
\usage{
\method{plot}{addHierBasis}(x, ind.func = 1, ind.lam = 1, ...)
}
\arguments{
\item{x}{An object of class \code{addHierBasis}.}

\item{ind.func}{The index of the component function to plot.}

\item{ind.lam}{The index of the lambda value to plot.}

\item{...}{Other arguments passed on to function \code{plot}.}
}
\description{
This function plots individual component functions for a specified value of
\code{lambda} for an object of class \code{addHierBasis}.
}
\examples{
require(Matrix)

set.seed(1)

# Generate the points x.
n <- 100
p <- 30

x <- matrix(rnorm(n*p), ncol = p)

# A simple model with 3 non-zero functions.
y <- rnorm(n, sd = 0.1) + sin(x[, 1]) + x[, 2] + (x[, 3])^3

mod <- AdditiveHierBasis(x, y, nbasis = 50, max.lambda = 30,
                         beta.mat = NULL,
                         nlam = 50, alpha = 0.5,
                         lam.min.ratio = 1e-4, m.const = 3,
                         max.iter = 300, tol = 1e-4)

# Plot the individual functions.
xs <- seq(-3,3,length = 300)
plot(mod,1,30, type  ="l",col = "red", lwd = 2, xlab = "x", ylab = "f_1(x)",
  main = "Estimating the Sine function")
lines(xs, sin(xs), type = "l", lwd = 2)
legend("topleft", c("Estimated Function", "True Function"),
      col = c("red", "black"), lwd = 2, lty = 1)

\dontrun{
plot(mod,2,30, type  ="l",col = "red", lwd = 2, xlab = "x", ylab = "f_2(x)",
  main = "Estimating the Linear function")
lines(xs, xs, type = "l", lwd = 2)
legend("topleft", c("Estimated Function", "True Function"),
      col = c("red", "black"), lwd = 2, lty = 1)

plot(mod,3,30, type  ="l",col = "red", lwd = 2, xlab = "x", ylab = "f_3(x)",
     main = "Estimating the cubic polynomial")
lines(xs, xs^3, type = "l", lwd = 2)
legend("topleft", c("Estimated Function", "True Function"),
       col = c("red", "black"), lwd = 2, lty = 1)
}



}
\references{
Haris, A., Shojaie, A. and Simon, N. (2018). Nonparametric Regression with
Adaptive Smoothness via a Convex Hierarchical Penalty. Available on request
by authors.
}
\seealso{
\code{\link{AdditiveHierBasis}}, \code{\link{predict.addHierBasis}}
}
\author{
Asad Haris (\email{aharis@uw.edu}),
Ali Shojaie and Noah Simon
}
