% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_basis_additive.R
\name{predict.addHierBasis}
\alias{predict.addHierBasis}
\title{Model Predictions for Additive HierBasis}
\usage{
\method{predict}{addHierBasis}(object, new.x = NULL, ...)
}
\arguments{
\item{object}{A fitted object of class '\code{addHierBasis}'.}

\item{new.x}{An optional matrix of values of \code{x} at which to predict.
The number of columns of \code{new.x} should be equal to the number of
columns of \code{object$x}.}

\item{...}{Not used. Other arguments for predict function.}
}
\value{
\item{fitted.values}{A matrix of fitted values with \code{nrow(new.x)}
                     rows and \code{nlam} columns}
}
\description{
The generic S3 method for predictions for objects of
class \code{addHierBasis}.
}
\details{
This function returns a matrix of  predicted values at the specified
values of x given by \code{new.x}. Each column corresponds to a lambda value
used for fitting the original model.

If \code{new.x == NULL} then this function simply returns
the fitted values of the estimated function at the original x values used for
model fitting. The predicted values are presented for each lambda values.
}
\examples{

library(HierBasis)
require(Matrix)

set.seed(1)

# Generate the points x.
n <- 100
p <- 30

x <- matrix(rnorm(n*p), ncol = p)

# A simple model with 3 non-zero functions.
y <- rnorm(n, sd = 0.1) + sin(x[, 1]) + x[, 2] + (x[, 3])^3

mod <- AdditiveHierBasis(x, y, nbasis = 50, max.lambda = 30,
                         beta.mat = NULL,
                         nlam = 50, alpha = 0.5,
                         lam.min.ratio = 1e-4, m.const = 3,
                         max.iter = 300, tol = 1e-4)

# Obtain predictions for new.x.
preds <- predict(mod, new.x = matrix(rnorm(n*p), ncol = p))

}
\author{
Asad Haris (\email{aharis@uw.edu}),
Ali Shojaie and Noah Simon
}
\references{
Haris, A., Shojaie, A. and Simon, N. (2016). Nonparametric Regression with
Adaptive Smoothness via a Convex Hierarchical Penalty. Available on request
by authors.
}
\seealso{
\code{\link{AdditiveHierBasis}}, \code{\link{plot.addHierBasis}}
}

