% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_basis_additive.R
\name{print.addHierBasis}
\alias{print.addHierBasis}
\title{Print the coefficients of fitted models.}
\usage{
\method{print}{addHierBasis}(x, lam.index = NULL, digits = 3, ...)
}
\arguments{
\item{x}{A fitted model, an object of class \code{addHierBasis}.}

\item{lam.index}{Specify the lambda indices to view.
If NULL then all lambda values are shown.}

\item{digits}{The significant figures printed.}

\item{...}{Not used, extra arguments to print.}
}
\description{
Print the coefficients of fitted models.
}
\examples{
\dontrun{
require(Matrix)

set.seed(1)

# Generate the points x.
n <- 100
p <- 30

x <- matrix(rnorm(n*p), ncol = p)

# A simple model with 3 non-zero functions.
y <- rnorm(n, sd = 0.1) + sin(x[, 1]) + x[, 2] + (x[, 3])^3

mod <- AdditiveHierBasis(x, y, nbasis = 50, max.lambda = 30)
print(mod, lam.index = 30)
}

}
\references{
Haris, A., Shojaie, A. and Simon, N. (2016). Nonparametric Regression with
Adaptive Smoothness via a Convex Hierarchical Penalty. Available on request
by authors.
}
\seealso{
\code{\link{AdditiveHierBasis}},
\code{\link{view.model.addHierBasis}}
}
\author{
Asad Haris (\email{aharis@uw.edu}),
Ali Shojaie and Noah Simon
}
