% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_basis_additive.R
\name{view.model.addHierBasis}
\alias{view.model.addHierBasis}
\title{View Fitted Additive Model}
\usage{
view.model.addHierBasis(object, lam.index = 1, digits = 3)
}
\arguments{
\item{object}{An object of class \code{addHierBasis}.}

\item{lam.index}{The index of the \code{lambda} value to view.}

\item{digits}{The significant figures printed.}
}
\description{
This function prints a list of size \code{p} where \code{p} is the number of
covariates. Each element in the list is a vector of strings showing the polynomial
representation of each individual component function.
}
\examples{
\dontrun{
require(Matrix)

set.seed(1)

# Generate the points x.
n <- 100
p <- 30

x <- matrix(rnorm(n*p), ncol = p)

# A simple model with 3 non-zero functions.
y <- rnorm(n, sd = 0.1) + sin(x[, 1]) + x[, 2] + (x[, 3])^3

mod <- AdditiveHierBasis(x, y, nbasis = 50, max.lambda = 30)
view.model.addHierBasis(mod, 30)
}
}
\author{
Asad Haris (\email{aharis@uw.edu}),
Ali Shojaie and Noah Simon
}
\references{
Haris, A., Shojaie, A. and Simon, N. (2016). Nonparametric Regression with
Adaptive Smoothness via a Convex Hierarchical Penalty. Available on request
by authors.
}
\seealso{
\code{\link{AdditiveHierBasis}}, \code{\link{print.addHierBasis}}
}

