% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_basis.R
\name{GetDoF.HierBasis}
\alias{GetDoF.HierBasis}
\title{Extract Degrees of Freedom}
\usage{
GetDoF.HierBasis(object)
}
\arguments{
\item{object}{An object of class \code{HierBasis}}
}
\value{
\item{dof}{A vector of degrees of freedom for the sequence of lambda values
            used for the fitted HierBasis model.}
}
\description{
Extract Degrees of Freedom
}
\examples{
require(Matrix)

set.seed(1)

# Generate the points x.
n <- 300
x <- (1:300)/300

# A simple quadratic function.
y1 <- 5 * (x - 0.5)^2
y1dat <- y1 + rnorm(n, sd = 0.1)

poly.fit <- HierBasis(x, y1dat)
dof <- GetDoF.HierBasis(poly.fit)

\donttest{
ind <- which.min(abs(dof - 3))[1]
plot(x, y1dat, type = "p", ylab = "y1")
lines(x, y1, lwd = 2)
lines(x, poly.fit$fitted.values[, ind], col = "red", lwd = 2)
}

}
\references{
Haris, A., Shojaie, A. and Simon, N. (2018). Nonparametric Regression with
Adaptive Smoothness via a Convex Hierarchical Penalty. Available on request
by authors.
}
\seealso{
\code{\link{HierBasis}}, \code{\link{predict.HierBasis}}
}
\author{
Asad Haris (\email{aharis@uw.edu}),
Ali Shojaie and Noah Simon
}
